be selective about which 3rd party jars to include when making jar files
Rather than including all of the third party jar files in every jar file
we make (warpdrive, readdb, etc), only include the ones we need.  This
is a little error prone in that forgetting to include a jar file will
cause runtime errors, but it makes for much smaller jar files for
distribution.