***Commiting code for Giorgos*** - Added a structure AggregatedData which takes the condition-specific data and aggregates them all. It stores the results in static variables to be accessible whenever needed.
- MultiIndependentMixtureCounts.train():
After the training, I added a small piece of code where I evaluate the log-likelihood of the ALL the data. I'm doing this because I'm using it in the BindingMixture to see if the components of a region (after an update) have converged.

- MultiIndependentMixtureCounts.serial_hard_train(), .serial_soft_train() and .parallel_train():
Now I'm forcing the algorithm to start checking for convergence after it passed <tt>maxAnnealIters</tt> number of iterations. This is not principled but it has been shown to give better results. That is, better sparsity.

- MultiIndependentMixtureCounts.convergence_check():
I renamed this method and made it public so as to be accessible by other classes as well. E.g. BindingMixture.

- I added a get_glob_loglik() method.

- Fixed a bug in the M_step() method.
Previously, I was giving sum_ga to prior_weight through the clone() which was really UNFORTUNATELY creating a local, temporary copy of prior_weight which was NOT passed up in the hierarchy. Now, this has been fixed.