active/inactive flag for analyses
changes to allow chipseq analyses in the DB to
be marked as active or inactive.  By default, the analysis selection in
WD will only show active analyses.  Also by default, my GPS script will
submit the final round of output as active and the previous rounds as
inactive.

There's a GUI application
edu.mit.csail.cgs.warpdrive.components.ChipSeqAnalysisFrame (you need to
call with a --species) that lets you set analyses as active or inactive,
so it should be easy to keep around old GPS (or other analysis) runs
without cluttering up the lists that we see in WD or dump to the
PPG/Syscode wikis.